data-enrichment/1-translation/pyproject.toml
[tool.poetry]
name = "xcurator-translation"
version = "0.0.0" # placeholder | version is defined in git tags | command: poetry version
description = "Translate artefact text data into the defined languages"
authors = ["Sören Räuchle <sraeuchle@3pc.de>"]
license = "MIT"
readme = "README.md"

# https://pypi.org/project/poetry-dynamic-versioning/
[tool.poetry-dynamic-versioning]
enable = false
vcs = 'git'
style = "semver"
pattern = '(?x)^v*((?P<epoch>\d+)!)?(?P<base>\d+(\.\d+)*)' # 0.0.0 || v0.0.0 (git tag)

[tool.poetry.build]
generate-setup-file = false

[tool.poetry.dependencies]
python = "3.12.0"
pandas = "2.1.1"
openpyxl = "3.1.2"
tables = "3.9.1"
httpx = "0.25.0"
cftime = "1.6.2"
more-itertools="10.1.0"
3pc-py-config = {path="./lib/3pc-py-config-1.1.1.tar.gz"}
beautifulsoup4 = "4.12.2"
tqdm = "4.66.1"
requests = "2.31.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.6"

[tool.poetry.group.test.dependencies]
pytest = "7.4.2"
pytest-asyncio = "0.21.0"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend" # wrapper around poetry.core.masonry.api

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"